+++ Examples/Example01.ps1
@@ -5,7 +5,8 @@
 [CmdletBinding()]
 param (
     [System.IO.FileInfo] $Path = '~\Desktop\',
-    [array] $Format = @('png')
+    [array] $Format = @('png'),
+    [bool] $DraftMode = $false
 )
 
 <#
@@ -45,6 +46,7 @@ $example1 = & {
     -MainDiagramLabel: Sets the diagram's title.
     -Filename: Specifies the output file name (without extension).
     -LogoName: Selects an image from the hashtable to use as the diagram logo.
+    -DraftMode: If set to $true, generates a draft version of the diagram for troubleshooting.
     If the specified logo image is not found, a default no_icon.png is used.
 #>
-New-Diagrammer -InputObject $example1 -OutputFolderPath $OutputFolderPath -Format $Format -MainDiagramLabel $MainGraphLabel -Filename Example1 -LogoName "Main_Logo"
+New-Diagrammer -InputObject $example1 -OutputFolderPath $OutputFolderPath -Format $Format -MainDiagramLabel $MainGraphLabel -Filename Example1 -LogoName "Main_Logo"  -DraftMode:$DraftMode