+++ Examples/Example15.ps1
@@ -1,4 +1,4 @@
-#    ** This time, we'll demonstrate the use of the Signature (Add-DiaHtmlSignatureTable) feature (Part of Diagrammer.Core module). **
+#    ** This example demonstrates how to use the Add-DiaNodeSpacer cmdlet to add spacer nodes that assist with diagram alignment. (part of Diagrammer.Core) **
 
 <#
     This example demonstrates how to create a 3-tier web application diagram using the Diagrammer module.
@@ -71,7 +71,7 @@ $DBServerInfo = [PSCustomObject][ordered]@{
     'Edition' = "Enterprise"
 }
 
-$example13 = & {
+$example15 = & {
     <#
         A SubGraph allows you to group objects in a container, creating a graph within a graph.
         SubGraph, Node, and Edge have attributes for setting background color, label, border color, style, etc.
@@ -100,12 +100,12 @@ $example13 = & {
         SubGraph Signature -Attributes @{Label = $Signature; fontsize = 22; penwidth = 1.5; labelloc = 'b'; labeljust = 'right'; style = "invis"; color = "darkgray" } {
 
             <#
-            The $WebServerFarm variable is an array of hashtables, each representing a web server node with its properties.
-            Each hashtable contains:
-            - Name: The name of the web server.
-            - AdditionalInfo: A custom object with properties to display in the node label.
-            - IconType: The type of icon to use for the node (must match a key in the $Images hashtable).
-        #>
+                The $WebServerFarm variable is an array of hashtables, each representing a web server node with its properties.
+                Each hashtable contains:
+                - Name: The name of the web server.
+                - AdditionalInfo: A custom object with properties to display in the node label.
+                - IconType: The type of icon to use for the node (must match a key in the $Images hashtable).
+            #>
 
             $WebServerFarm = @(
                 @{
@@ -141,16 +141,12 @@ $example13 = & {
             )
 
             <#
-            This time, we will simulate a Web Server Farm with multiple web server node. While the Add-DiaNodeIcon cmdlet is typically used to add icons/properties to nodes, it lack the ability to create multiple nodes with distinct properties.
+                This time, we will simulate a Web Server Farm with multiple web server node. While the Add-DiaNodeIcon cmdlet is typically used to add icons/properties to nodes, it lack the ability to create multiple nodes with distinct properties.
 
-            Add-DiaHTMLNodeTable has the capability to create a table layout for the nodes simulting a web server farm. It also allows the addition of icons and properties to each node in the table.
+                Add-DiaHTMLNodeTable has the capability to create a table layout for the nodes simulting a web server farm. It also allows the addition of icons and properties to each node in the table.
 
-            ** The $Images object and IconType "Server" must be defined earlier in the script **
-
-            -MultiIcon parameter allows multiple icons to be displayed in the table. (IconType must be specified in the inputObject)
-            -iconType parameter sets the type of icon to use for the nodes. In this case the $WebServerFarm.IconType hashtable value is used
-            (must match a key in the $Images hashtable).
-        #>
+                ** The $Images object and IconType "Server" must be defined earlier in the script **
+            #>
 
             $Web01Label = Add-DiaHTMLNodeTable -ImagesObj $Images -inputObject $WebServerFarm.Name -iconType $WebServerFarm.IconType -columnSize 3 -AditionalInfo $WebServerFarm.AdditionalInfo -Subgraph -SubgraphLabel "Web Server Farm" -SubgraphLabelPos "top" -SubgraphTableStyle "dashed,rounded" -TableBorderColor "gray" -TableBorder "1" -SubgraphLabelFontsize 20 -fontSize 18 -MultiIcon -DraftMode:$DraftMode
 
@@ -163,23 +159,23 @@ $example13 = & {
             Node -Name DB01 -Attributes @{Label = $DB01Label; shape = 'plain'; fillColor = 'transparent'; fontsize = 14 }
 
             <#
-            This section creates connections between the nodes in a hierarchical layout.
-            The Edge statements create connections between the nodes. (Edge is a reserved word in the PSGraph module)
-            https://psgraph.readthedocs.io/en/latest/Command-Edge/
-        #>
+                This section creates connections between the nodes in a hierarchical layout.
+                The Edge statements create connections between the nodes. (Edge is a reserved word in the PSGraph module)
+                https://psgraph.readthedocs.io/en/latest/Command-Edge/
+            #>
 
             Edge -From Web01 -To App01 @{label = 'gRPC'; color = 'black'; fontsize = 14; fontcolor = 'black'; minlen = 3 }
             Edge -From App01 -To DB01 @{label = 'SQL'; color = 'black'; fontsize = 14; fontcolor = 'black'; minlen = 3 }
 
             <#
-            The Rank cmdlet is used to place nodes at the same hierarchical level.
-            In this example, App01 and DB01 are aligned horizontally.
-        #>
+                The Rank cmdlet is used to place nodes at the same hierarchical level.
+                In this example, App01 and DB01 are aligned horizontally.
+            #>
             Rank -Nodes App01, DB01
 
             <#
-            In this section, we add a network router and a cloud icon to represent internet connectivity.
-        #>
+                In this section, we add a network router and a cloud icon to represent internet connectivity.
+            #>
 
             $RouterInfo = [PSCustomObject][ordered]@{
                 'OS' = 'Cisco IOS'
@@ -193,21 +189,19 @@ $example13 = & {
             Edge -From Router01 -To Web01 @{label = 'GE0/0'; color = 'black'; fontsize = 18; fontcolor = 'black'; minlen = 2 }
 
             <#
-            This section demonstrates the use of the Add-DiaNodeImage to add a custom image to the diagram (Part of Diagrammer.Core module).
+                This section demonstrates the use of the Add-DiaNodeImage to add a custom image to the diagram (Part of Diagrammer.Core module).
 
-            -Name parameter sets the name of the node (WAN in this case).
-            -ImageSizePercent parameter sets the size of the image as a percentage (30% in this case).
-        #>
+                -Name parameter sets the name of the node (WAN in this case).
+                -ImageSizePercent parameter sets the size of the image as a percentage (30% in this case).
+            #>
 
             Add-DiaNodeImage -Name "WAN" -ImagesObj $Images -IconType "Cloud" -IconPath $IconPath -ImageSizePercent 30 -NodeObject -DraftMode:$DraftMode
 
-            Edge -From WAN -To Router01 @{label = 'Serial0/0'; color = 'black'; fontsize = 18; fontcolor = 'black'; minlen = 2 }
-
             <#
-            The Add-DiaHTMLTable cmdlet is used to create a list table to display additional information about a object.
+                The Add-DiaHTMLTable cmdlet is used to create a list table to display additional information about a object.
 
-           In this example, we create a table to display the router's network interface information.
-        #>
+                In this example, we create a table to display the router's network interface information.
+            #>
 
             $RouterNetworkInfo = @(
                 "S0/0:"
@@ -221,17 +215,52 @@ $example13 = & {
             Edge -From Router01 -To RouterNetworkInfo @{color = 'black'; fontsize = 18; fontcolor = 'black'; minlen = 1; style = 'filled'; arrowhead = 'none'; arrowtail = 'none' }
 
             Rank Router01, RouterNetworkInfo
+
+            <#
+                The Add-DiaNodeShape cmdlet is used to create a node with a specific shape and various customizable attributes (Part of Diagrammer.Core module). Supported Shapes: https://graphviz.org/doc/info/shapes.html
+
+                ** In this example, we create a rectangle to simulate a firewall presence in the network. **
+            #>
+
+            Add-DiaNodeShape -Name "Firewall" -Shape rectangle -ShapeStyle 'filled' -ShapeFillColor 'red:white' -ShapeFontSize 14 -ShapeFontColor 'black' -ShapeFontName 'Arial' -ShapeWidth 3 -ShapeLabelPosition center -ShapeLineColor 'black' -DraftMode:$DraftMode
+
+            Edge -From WAN -To Firewall @{labeldistance = 5; headlabel = 'port1'; color = 'black'; fontsize = 18; fontcolor = 'black'; minlen = 2; arrowhead = 'normal'; arrowtail = 'normal' }
+            Edge -From Firewall -To Router01 @{labeldistance = 4; headlabel = 'Serial0/0'; taillabel = 'port2'; color = 'black'; fontsize = 18; fontcolor = 'black'; minlen = 2; arrowhead = 'normal'; arrowtail = 'normal' }
+
+
+            <#
+                The Add-DiaNodeSpacer cmdlet is used to create invisible spacer nodes that help with diagram alignment (Part of Diagrammer.Core module).
+
+                In this diagram there is a aliment issue with the Web01 node, as it is not centered with the App01 and DB01 nodes below it. To fix this,
+                we create two spacer nodes (FillerRight and FillerLeft) on either side of the Web01 node.
+
+                -Name parameter sets the name of the spacer node.
+                -ShapeWidth parameter sets the width of the spacer node.
+                -ShapeHeight parameter sets the height of the spacer node.
+                -ShapeOrientation parameter sets the orientation of the spacer node (0 for horizontal, 1 for vertical).
+                -DraftMode parameter specifies whether the node is in draft mode.
+
+                The rank command is used to align the spacer nodes and the web server node horizontally.
+                The edge commands create invisible edges between the spacer nodes and the web server node to maintain alignment.
+
+                ** In this example, we create two spacer nodes to help align the web server node. **
+            #>
+
+            Add-DiaNodeSpacer -Name "FillerRight" -ShapeWidth 2 -ShapeHeight 1 -ShapeOrientation 0 -DraftMode:$DraftMode
+            Add-DiaNodeSpacer -Name "FillerLeft" -ShapeWidth 2 -ShapeHeight 1 -ShapeOrientation 0 -DraftMode:$DraftMode
+
+            Rank -Nodes FillerLeft, Web01, FillerRight
+
+            $Style = if ($DraftMode) { 'filled' } else { 'invis' }
+            $StyleColor = if ($DraftMode) { 'red' } else { 'transparent' }
+
+            Edge -From FillerLeft -To Web01 @{minlen = 2; style = $Style; color = $StyleColor; }
+            Edge -From Web01 -To FillerRight @{minlen = 2; style = $Style; color = $StyleColor; }
         }
     }
 }
-
 <#
     This command generates the diagram using the New-Diagrammer cmdlet (part of Diagrammer.Core).
-
-    ** The WaterMark feature is demonstrated here, which adds a watermark text to the diagram. **
-
-    -WaterMarkText parameter sets the text for the watermark.
-    -WaterMarkColor parameter sets the color of the watermark text.
 #>
 
-New-Diagrammer -InputObject $example13 -OutputFolderPath $OutputFolderPath -Format $Format -MainDiagramLabel $MainGraphLabel -Filename Example13 -LogoName "Main_Logo" -Direction top-to-bottom -IconPath $IconPath -ImagesObj $Images -WaterMarkText "Confidential" -WaterMarkColor "DarkGray" -WaterMarkFontOpacity 40 -DraftMode:$DraftMode
+New-Diagrammer -InputObject $example15 -OutputFolderPath $OutputFolderPath -Format $Format -MainDiagramLabel $MainGraphLabel -Filename Example15 -LogoName "Main_Logo" -Direction top-to-bottom -IconPath $IconPath -ImagesObj $Images -DraftMode:$DraftMode